Put all dust-related codes into a separated file
